feat: Make newline forced newline and add space for soft newlines
In some instances it can be necessary for a newline to appear regardless of whether a `Doc` fits on a line. Since newline seems more appropriate for this behavior I think its appropriate to do this breaking change.

BREAKING CHANGE

`newline` now always inserts `\n`, to get the old behavior use `space` instead