fix: Newlines should not cause the rest to be written on a single line
If `fitting` encounters a `NewLine` it will immediately exit since the documents before `Newline` will fit the line. The documents after that `Newline` may however not fit but will still get `Mode::Flat` assigned since there is no way to know if they are before or after that newline so this commit recalculates `fitting` on the remaining documents.