Add quick start documentation. Fixes #26.
This is by no means complete, but it should help get a new user up and
running quickly.  Please feel free to edit for tone and content.

Locally, the "Quick start" and "Advanced usage" headers do not render
correctly, but I see other examples with this syntax that do render
correctly on crates.io.  I can't explain this.